Move AppManifests to separate file.
